Use Glide

+++ app/src/main/java/com/example/proyectofinalv2/domain/model/Note.kt
@@ -3,12 +3,11 @@ package com.example.proyectofinalv2.domain.model
 import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.PrimaryKey
-import java.time.LocalDate
 import java.util.*
 
 @Entity
 data class Note(
-    @PrimaryKey(autoGenerate = true) @ColumnInfo(name = "id") val id: Int = 0,
+    @PrimaryKey(autoGenerate = true) @ColumnInfo(name = "id") val id: Long = 0,
     @ColumnInfo(name = "title") var title: String?,
     @ColumnInfo(name = "description") var description: String?,
     @ColumnInfo(name = "isTask") var isTask: Boolean?,
